- Version increased to 0.8.9 - Money_t.h: changed 'include <xxx>' to 'include "xxx"' for a local, project include file 'DecVal.h' - Database.cpp: added casting to an int to avoid implicit conversion - Database.cpp: corrected conversions from DB types in the method "Database::commodities()" - Database: method "Database::commodities()" now accepts "long long" type instead of "qint64". The change is now consistent with "toLongLong()" method of QVariant - CurrencyData.h: added inclusion of QObject file. It's necessary for trUtf8() methods - InvoiceDialog.ui: changing text of "labelDiscount2" from "Rabat:" to "Rabat brutto:" for better clarity -InvoiceDialogImpl.cpp: removed incorrect references to "currentIndex" in the mapper - Counterparty_t: hiding class fields and introducing getters and setters. Correcting references to Counterparty_t in other files - GuiUserEditInvoice: introducing methods for passing numbers about total sums - InvoiceSums: introducing class so that repeating things done in tst_invoicedialogtest.cpp and tst_mainwindowtest.cpp can be uniqued - tst_invoicedialogtest.cpp: ordering and extending unit tests - tst_mainwindowtest.cpp: ordering and extending unit tests

+++ qfaktury/src/model/Database.h
@@ -138,7 +138,7 @@ public:
      * @param id_invoice
      * @return QList<CommodityVisualData>
      */
-    QList<CommodityVisualData> commodities(const qint64 id_invoice);
+    QList<CommodityVisualData> commodities(const long long id_invoice);
     /**
      * @brief
      *
